Conflicts:
	examples/itemviews/frozencolumn/main.cpp
	src/declarative/canvas/qsimplecanvas.cpp
	src/declarative/canvas/qsimplecanvas_p.h
	src/declarative/canvas/qsimplecanvasitem.h
	src/declarative/extra/qfxparticles.cpp
	src/declarative/fx/fx.pri
	src/declarative/fx/qfxblurfilter.h
	src/declarative/fx/qfxcontentwrapper.cpp
	src/declarative/fx/qfxflickable.cpp
	src/declarative/fx/qfxfocuspanel.h
	src/declarative/fx/qfxfocusrealm.h
	src/declarative/fx/qfxhighlightfilter.cpp
	src/declarative/fx/qfxhighlightfilter.h
	src/declarative/fx/qfximage.cpp
	src/declarative/fx/qfxitem.cpp
	src/declarative/fx/qfxitem.h
	src/declarative/fx/qfxrect.cpp
	src/declarative/fx/qfxreflectionfilter.h
	src/declarative/fx/qfxshadowfilter.cpp
	src/declarative/fx/qfxshadowfilter.h
	src/declarative/fx/qfxtext.cpp
	src/declarative/fx/qfxtext.h
	src/declarative/fx/qfxtextedit.cpp
	src/declarative/opengl/glbasicshaders.h
	src/declarative/test/qfxtestengine.cpp
	src/declarative/test/qfxtestengine.h
	src/declarative/test/qfxtestobjects.cpp
	src/declarative/test/qfxtestobjects.h
	src/declarative/test/qfxtestview.h
	src/declarative/util/qfxglobal.h
	src/declarative/util/qfxview.cpp
	src/gui/graphicsview/qgraphicsitem_p.h
	tools/qmlviewer/qmlviewer.cpp

+++ src/declarative/fx/qfxitem_p.h
@@ -143,7 +143,7 @@ public:
     QList<QUrl> _qmlnewloading;
     QList<QmlComponent*> _qmlnewcomp;
 
-    QmlNullableValue<int> _baselineOffset;
+    QmlNullableValue<qreal> _baselineOffset;
     float _rotation;
 
     bool _classComplete:1;
